refactor: export vector store only in nodejs environment on top level
