+++ packages/llamaindex/src/index.ts
@@ -17,3 +17,7 @@ export { GeminiVertexSession } from "./llm/gemini/vertex.js";
 // Expose AzureDynamicSessionTool for node.js runtime only
 export { JinaAIEmbedding } from "./embeddings/JinaAIEmbedding.js";
 export { AzureDynamicSessionTool } from "./tools/AzureDynamicSessionTool.node.js";
+
+// Don't export vector store modules for non-node.js runtime on top level,
+//  as we cannot guarantee that they will work in other environments
+export * from "./vector-store.js";
