+++ packages/llamaindex/src/vector-store.ts
@@ -1,0 +1,1 @@
+export * from "./vector-store/index.js";
