+++ packages/llamaindex/src/vector-store/WeaviateVectorStore.ts
